Create pipeline process subclass for remote pipeline runtimes
- new subclass RuntimePipelineProcess added for kfp and airflow processor
- fix application package to include kfp schema
