+++ packages/services/src/test/application.spec.ts
@@ -100,7 +100,12 @@ describe('@elyra/services', () => {
         const schemaNames = schemas.map((schema: any) => {
           return schema.name;
         });
-        const knownSchemaNames = ['code-snippet', 'runtime-image', 'airflow'];
+        const knownSchemaNames = [
+          'code-snippet',
+          'runtime-image',
+          'kfp',
+          'airflow'
+        ];
         for (const schemaName of knownSchemaNames) {
           expect(schemaNames).toContain(schemaName);
         }
